move contract class tests to its one file
