+++ crates/papyrus_gateway/src/gateway_test.rs
@@ -1,7 +1,6 @@
 use std::panic;
 
 use assert_matches::assert_matches;
-use camelpaste::paste;
 use futures_util::future::join_all;
 use hyper::{header, Body, Request};
 use jsonrpsee::core::client::ClientT;
@@ -14,18 +13,11 @@ use papyrus_storage::header::HeaderStorageWriter;
 use papyrus_storage::test_utils::get_test_storage;
 use pretty_assertions::assert_eq;
 use starknet_api::block::{BlockHash, BlockHeader, BlockNumber, BlockStatus};
-use starknet_api::deprecated_contract_class::{
-    EventAbiEntry, FunctionAbiEntryWithType, StructAbiEntry,
-};
-use test_utils::{get_rng, GetTestInstance};
 use tower::BoxError;
 
 use crate::api::JsonRpcError;
 use crate::middleware::proxy_rpc_request;
 use crate::test_utils::get_test_gateway_config;
-use crate::v0_3_0::deprecated_contract_class::{
-    ContractClassAbiEntryType, ContractClassAbiEntryWithType,
-};
 use crate::version_config::VERSION_CONFIG;
 use crate::{get_block_status, run_server, SERVER_MAX_BODY_SIZE};
 
@@ -141,68 +133,3 @@ fn get_block_status_test() {
     assert_eq!(get_block_status(&txn, BlockNumber(1)).unwrap(), BlockStatus::AcceptedOnL2);
     assert_eq!(get_block_status(&txn, BlockNumber(2)).unwrap(), BlockStatus::AcceptedOnL2);
 }
-
-#[tokio::test]
-async fn test_contractclassabientrywithtype_from_api_contractclassabientry() {
-    let mut rng = get_rng();
-    let _: ContractClassAbiEntryWithType =
-        starknet_api::deprecated_contract_class::ContractClassAbiEntry::Event(
-            EventAbiEntry::get_test_instance(&mut rng),
-        )
-        .try_into()
-        .unwrap();
-    let _: ContractClassAbiEntryWithType =
-        starknet_api::deprecated_contract_class::ContractClassAbiEntry::Function(
-            FunctionAbiEntryWithType::get_test_instance(&mut rng),
-        )
-        .try_into()
-        .unwrap();
-    let _: ContractClassAbiEntryWithType =
-        starknet_api::deprecated_contract_class::ContractClassAbiEntry::Struct(
-            StructAbiEntry::get_test_instance(&mut rng),
-        )
-        .try_into()
-        .unwrap();
-}
-
-macro_rules! test_ContractClassAbiEntryType_from_FunctionAbiEntryType {
-    ($variant:ident) => {
-        paste! {
-            #[tokio::test]
-            #[allow(non_snake_case)]
-            async fn [< ContractClassAbiEntryType_from_FunctionAbiEntryType_ $variant:lower>]() {
-                let _: ContractClassAbiEntryType =
-                starknet_api::deprecated_contract_class::FunctionAbiEntryType::$variant
-                    .try_into()
-                    .unwrap();
-            }
-        }
-    };
-}
-test_ContractClassAbiEntryType_from_FunctionAbiEntryType!(Constructor);
-test_ContractClassAbiEntryType_from_FunctionAbiEntryType!(L1Handler);
-test_ContractClassAbiEntryType_from_FunctionAbiEntryType!(Function);
-
-// macro to generate a test that creates a ContractClassAbiEntry with a variant based on the given
-// variant input and call try_into().unwrap()
-macro_rules! test_contract_class_abi_entry_with_type {
-    ($variant:ident, $variant_inner:ident) => {
-        paste! {
-            #[tokio::test]
-            #[allow(non_snake_case)]
-            async fn [<ContractClassAbiEntryWithType_from_api_ContractClassAbiEntry_ $variant:lower>]() {
-                let mut rng = get_rng();
-                let _: ContractClassAbiEntryWithType =
-                    starknet_api::deprecated_contract_class::ContractClassAbiEntry::$variant(
-                        $variant_inner::get_test_instance(&mut rng),
-                    )
-                    .try_into()
-                    .unwrap();
-            }
-        }
-    };
-}
-
-test_contract_class_abi_entry_with_type!(Event, EventAbiEntry);
-test_contract_class_abi_entry_with_type!(Function, FunctionAbiEntryWithType);
-test_contract_class_abi_entry_with_type!(Struct, StructAbiEntry);
