+++ crates/papyrus_gateway/src/v0_3_0/mod.rs
@@ -1,6 +1,8 @@
 pub mod api;
 pub mod block;
 pub mod deprecated_contract_class;
+#[cfg(test)]
+mod deprecated_contract_class_test;
 pub mod state;
 pub mod transaction;
 #[cfg(test)]
